add redis client
handle eslint
