+++ controllers/FilesController.js
@@ -1,0 +1,7 @@
+const redisClient = require('../utils/redis');
+const dbClient = require('../utils/db');
+const postUpload = (req, res)=>{
+	const clientCollection = 
+
+
+}
