+++ utils/redis.js
@@ -13,12 +13,7 @@ class RedisClient {
   }
 
   isAlive() {
-    try {
-      return this.client.connected;
-    } catch (error) {
-      console.error(error);
-      return false;
-    }
+    return this.client.connected;
   }
 
   async get(key) {
@@ -30,31 +25,15 @@ class RedisClient {
   async set(key, value, expDuration) {
     const setAsync = promisify(this.client.set).bind(this.client);
     const expireAsync = promisify(this.client.expire).bind(this.client);
-
-    try {
-      await setAsync(key, value);
-      await expireAsync(key, expDuration);
-      // console.log(`${key}: ${value} set`);
-    } catch (error) {
-      console.error(error);
-    }
+    await setAsync(key, value);
+    await expireAsync(key, expDuration);
+    // console.log(`${key}: ${value} set`);
   }
 
   async del(key) {
-    const getAsync = promisify(this.client.get).bind(this.client);
     const delAsync = promisify(this.client.del).bind(this.client);
-
-    try {
-      const value = await getAsync(key);
-      if (value !== null) {
-        await delAsync(key);
-        console.log(`${key} deleted`);
-      } else {
-        console.log(`${key} does not exist`);
-      }
-    } catch (error) {
-      console.error(error);
-    }
+    await delAsync(key);
+    console.log(`${key} deleted`);
   }
 }
